requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
distlib==0.3.1
Django==3.1.6
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.6.4
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
learn==1.0.0
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
pandas==1.0.5
Pillow==8.0.1
psutil==5.7.0
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
six==1.15.0
sk==0.0.1
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.1
urllib3==1.25.9
virtualenv==20.4.2
wrapt==1.12.1
